requirements.txt
bleach==4.1.0
certifi==2020.12.5
charset-normalizer==2.0.4
colorama==0.4.4
distlib==0.3.1
Django==3.2.7
docutils==0.17.1
factory-boy==3.2.0
Faker==8.13.2
filelock==3.0.12
greenlet==1.0.0
idna==3.2
importlib-metadata==4.8.1
keyring==23.2.1
packaging==21.0
pipenv==2020.11.15
pkginfo==1.7.1
playwright==1.8.0a1
protobuf==3.14.0
pycodestyle==2.6.0
pyee==8.2.2
Pygments==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
readme-renderer==29.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.15.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.62.2
twine==3.4.2
typing-extensions==3.10.0.2
urllib3==1.26.6
virtualenv==20.4.3
virtualenv-clone==0.5.4
webencodings==0.5.1
zipp==3.5.0
